service/requirements-arc.txt


basicsr==1.4.2
compel==2.0.2
datasets==2.18.0
diffusers==0.27.2
Flask==3.0.3
huggingface_hub==0.22.2
ipex_llm==2.1.0b20240705
langchain==0.2.2
langchain_community==0.2.3
langchain_core==0.2.4
numpy==1.26.4
opencv_python==4.9.0.80
peft==0.10.0
Pillow==10.3.0
psutil==5.9.8
Requests==2.32.3
sentence_transformers==2.3.1
transformers==4.39.0
trl==0.9.3
bigdl-core-xe-21==2.5.0b20240705
bigdl-core-xe-addons-21==2.5.0b20240705
bigdl-core-xe-batch-21==2.5.0b20240705
faiss-cpu==1.8.0
docx2txt==0.8
pypdf==4.2.0
unstructured==0.14.6
setuptools==69.5.1
